create enum.ts and fix TypeScript errors
