+++ packages/step-checker/src/enums.ts
@@ -1,0 +1,11 @@
+export enum MistakeId {
+    EQN_ADD_DIFF,
+    EQN_MUL_DIFF,
+    EXPR_ADD_NON_IDENTITY,
+    EXPR_MUL_NON_IDENTITY,
+}
+
+export enum Status {
+    Correct,
+    Incorrect,
+}
