+++ packages/step-checker/src/index.ts
@@ -1,7 +1,8 @@
 import StepChecker from "./step-checker";
+
 export {checkStep} from "./step-checker";
+export {MistakeId} from "./enums";
+
 export type {Context, Mistake} from "./types";
-// NOTE: enums are types even though they can be used within types
-export {MistakeId} from "./types";
 
 export default StepChecker;
